feat(imbas): implement context and map block builders for intent injection
- Introduced `buildCtxBlock` and `buildMapBlock` functions to construct context and map injection text for directory visits.
- Refactored existing code to utilize these new utility functions, enhancing clarity and maintainability.
- Removed obsolete code related to context building from `intent-injector.ts` to streamline functionality.
